[Enhancement] Add file_utils module and improved function calling
