+++ wattelse/evaluation_pipeline/utils/__init__.py
@@ -3,7 +3,9 @@ Utility modules for the evaluation pipeline.
 """
 
 from wattelse.evaluation_pipeline.utils.port_manager import PortManager
+from wattelse.evaluation_pipeline.utils.file_utils import handle_output_path
 
 __all__ = [
     "PortManager",
+    "handle_output_path",
 ]
